+++ src/operations.py
@@ -18,6 +18,69 @@ COLORS = {
 }
 
 
+PRICES_REGEXP = {
+    "magazine": r"(.+)\sà vista",
+    "amazon": r"(.+)",
+    "americanas": r"(.+)",
+}
+
+
+def search_product(search_term_input, shops_list, shops_picker_msg, CONFIG=None):
+    shops_picker_msg = show_shops(CONFIG)
+
+    options = Options()
+    options.add_argument("--disable-extensions")
+    options.add_argument("--disable-gpu")
+    options.add_argument("--disable-infobars")
+
+    prefs = {"profile.managed_default_content_settings.images": 2}
+
+    options.add_experimental_option("prefs", prefs)
+    options.headless = True
+
+    shops = rx.from_(shops_list).pipe(
+        ops.map(
+            lambda shop: re.findall(f"\[{shop}\]\s(\w+)", shops_picker_msg)[0]
+        ),
+        ops.map(
+            lambda shop: {
+                **CONFIG.get(shop),
+                "priceRegexp": PRICES_REGEXP.get(shop),
+            }
+        ),
+        ops.filter(lambda el: el),
+    )
+
+    search_term = shops.pipe(
+        ops.pluck("innerChar"),
+        ops.map(lambda inner_char: inner_char.join(search_term_input.split())),
+    )
+
+    found_products = shops.pipe(
+        ops.zip(search_term),
+        ops.flat_map(lambda el: get_products(*el, options)),
+        ops.publish(),
+    )
+
+    return found_products
+
+
+def persist_prices_obs(found_products_obs):
+    persist_prices_obs = found_products_obs.pipe(
+        ops.to_list(),
+        ops.do_action(
+            lambda products: persist_prices([{search_term_input: products}])
+        ),
+    )
+    return persist_prices_obs
+
+
+def shop_picker_input(shops_picker_msg):
+    return input(
+        f"> Choose what shop(s) you want to search by typing the respective number(s):\n{shops_picker_msg}\n>> "
+    )
+
+
 def launch_browser(
     url: str, options: Options, config: Dict[str, str]
 ) -> webdriver:
