+++ test/src/test_operations.py
@@ -4,6 +4,10 @@ from unittest.mock import Mock
 from src.operations import show_shops, get_property
 
 
+with open('./config.json', 'r') as shops_config:
+    config = json.load(shops_config)
+
+
 mocked_parent = Mock()
 mocked_child = Mock()
 
@@ -29,3 +33,10 @@ class Main(unittest.TestCase):
             show_shops(config_mock),
             r"\[\d+\]\s\w+\n\[\d+\]\s\w+\n\[\d+\]\s\w+",
         )
+
+    def test_should_return_no_empty_observable(self):
+        for id, shop_name in enumerate(config):
+            shop_picker_input_mock = Mock(return_value=f'{id}')
+            products = search_product(search_term_input_mock, shop_picker_input_mock)
+            products.subscribe(lambda res: self.assertGreater(len(res), 0, msg=f'{shop_name} return nothing'))
+            products.connect()
